backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
alembic
psycopg2-binary
langchain
openai
httpx
gql
langchain-openai
langchain_community
fastapi-cors
python-jose
passlib
bcrypt
pydantic[email]
passlib==1.7.4
bcrypt==3.2.2
